ISIS-1781 fix wrong type: expected long, but is int

+++ core/viewer-restfulobjects-rendering/src/main/java/org/apache/isis/viewer/restfulobjects/rendering/domainobjects/JsonValueEncoder.java
@@ -248,7 +248,7 @@ public final class JsonValueEncoder {
                     return adapterFor(repr.asLong());
                 }
                 if (repr.isInt()) {
-                    return adapterFor(repr.asInt());
+                    return adapterFor(repr.asLong());
                 }
                 if (repr.isBigInteger()) {
                     return adapterFor(repr.asBigInteger().longValue());
